Update with working version rewritten today.
Reused the bs4 code but rewrote almost the entire rest of script from scratch. Instead of downloading the county list from the API, it is now already saved in file counties.json to be parsed. I toyed with using the Google Sheets API to input values directly into an online spreadsheet, but creating a csv from scratch is probably faster and definitely uses less bandwidth.